Provide support for remapping file descriptors.
When using Argdata as a format for RPCs across UNIX sockets, file
descriptor numbers embedded in the serialized data need to be remapped
to process-local file descriptors on the receiving side.

Extend argdata_from_buffer() to take a callback function that is invoked
every time we need to obtain a process-local file descriptor number.
This function can then be used by argdata_get_fd() and
argdata_serialize() to transform nodes when accessed.